requirements.txt
suds
tornado==5.0
peewee==3.2
PyMySQL==0.8
bcrypt==3.1
redis==2.10
rsa==3.4
pillow==5.1
rq==0.10
requests==2.18
xlrd==1.1
xlwt==1.3
qrcode==6.0